requirements.txt
torch
torchvision
rl-games
gym-super-mario-bros
opencv-python
tensorboard
ray
numpy
PyYAML
